[src] split src code to core and lib
