+++ src/core/omi.js
