+++ src/core/style.js
