+++ src/index.js
@@ -1,7 +1,8 @@
-import Omi from './omi.js'
-import soda from './soda.js'
-import Component from './component.js'
-import Store from './store.js'
+import Omi from './core/omi.js'
+import Store from './core/store.js'
+import Component from './core/component.js'
+
+import soda from './lib/soda.js'
 
 soda.prefix('o')
