+++ src/index.lite.js
@@ -1,6 +1,6 @@
-import Omi from './omi.js'
-import Component from './component.js'
-import Store from './store.js'
+import Omi from './core/omi.js'
+import Store from './core/store.js'
+import Component from './core/component.js'
 
 Omi.template = function(tpl, data) {
     return tpl
