+++ src/lib/morphdom.js
