--- test/js/hello.js
@@ -1,10 +1,0 @@
-
-describe('A suite is just a function', function() {
-    var a
-
-    it('and so is a spec', function() {
-        a = true
-
-        expect(a).toBe(true)
-    })
-})
