+++ test/js/style.js
@@ -1,5 +1,5 @@
 import assert from 'assert'
-import style from '../../src/style.js'
+import style from '../../src/core/style.js'
 
 describe('scoper', function() {
     describe('Attribute selector as prefix', function() {
